requirements.txt
PySide6>=6.6
numpy
Pillow
ultralytics>=8.2
faiss-cpu
